[#80196712] Adds glesUtil namespace
One of the issue in previous implementation is that when splitting large
geometries, we used osgUtil::VertexAccessOrderVisitor to renumber vertices.
The problem is that the current osg implementation skips POINTS and LINES
primitives.

To regroup all code that is mostly duplicated from osg (e.g. part of osgUtil
code is defined in the MeshOptimizers.cpp thus not reusable), we therefore
introduce the glesUtil namespace.