[#95964390] Updates AnimationVisitor
We now bake influence maps in vertex attribs.

As a first step, we look for RigGeometry skeleton using the same code as in
`osgAnimation::UpdateRigGeometry`. We only keep the 4 bones/weights with largest influence and
remap RigGeometry bones locally:
* use an index local to the geometry
* use uservalues to map local index to bone name

Note that the AnimationVisitor re-adds the vertex attribs to add a specific user values (it does not
cost more memory).