Simplifies json values serialization
Some `JSONValue` template specialization are now handled by `json_stream
directly so there is no need to duplicate e.g. `NaN` filtering.

We still need a `JSONValue<std::string>` specialization as `json_stream`
doesn't check wether the string is quoted or not (and it simpler to keep
things like this for now).